backend/requirements.txt
fastapi 
pydantic
python-dotenv
httpx